Extract modules.
