+++ lib/main.ts
@@ -14,6 +14,7 @@ import base = require('./base');
 import CallTranspiler = require('./call');
 import DeclarationTranspiler = require('./declaration');
 import ExpressionTranspiler = require('./expression');
+import ModuleTranspiler = require('./module');
 import StatementTranspiler = require('./statement');
 
 export interface TranspilerOptions {
@@ -45,6 +46,7 @@ export class Transpiler {
       new CallTranspiler(this),
       new DeclarationTranspiler(this),
       new ExpressionTranspiler(this),
+      new ModuleTranspiler(this, options.generateLibraryName),
       new StatementTranspiler(this),
     ];
   }
@@ -173,43 +175,6 @@ export class Transpiler {
     });
   }
 
-  private visitExternalModuleReferenceExpr(expr: ts.Expression) {
-    // TODO: what if this isn't a string literal?
-    var moduleName = <ts.StringLiteralExpression>expr;
-    var text = moduleName.text;
-    if (text.match(/^\.\//)) {
-      // Strip './' to be more Dart-idiomatic.
-      text = text.substring(2);
-    } else if (!text.match(/^\.\.\//)) {
-      // Unprefixed imports are package imports.
-      text = 'package:' + text;
-    }
-    moduleName.text = text + '.dart';
-    this.visit(expr);
-  }
-
-  private static isIgnoredAnnotation(e: ts.ImportSpecifier) {
-    var name = base.ident(e.name);
-    switch (name) {
-      case 'CONST':
-      case 'ABSTRACT':
-      case 'IMPLEMENTS':
-        return true;
-      default:
-        return false;
-    }
-  }
-
-  private isEmptyImport(n: ts.ImportDeclaration): boolean {
-    var bindings = n.importClause.namedBindings;
-    if (bindings.kind != ts.SyntaxKind.NamedImports) return false;
-    return (<ts.NamedImports>bindings).elements.every(Transpiler.isIgnoredAnnotation);
-  }
-
-  private filterImports(ns: ts.ImportOrExportSpecifier[]) {
-    return ns.filter((e) => !Transpiler.isIgnoredAnnotation(e));
-  }
-
   hasFlag(n: {flags: number}, flag: ts.NodeFlags): boolean {
     return n && (n.flags & flag) !== 0 || false;
   }
@@ -226,7 +191,7 @@ export class Transpiler {
     'StringMap': 'Map'
   };
 
-  private visitTypeName(typeName: ts.EntityName) {
+  visitTypeName(typeName: ts.EntityName) {
     if (typeName.kind !== ts.SyntaxKind.Identifier) {
       this.visit(typeName);
       return;
@@ -250,7 +215,7 @@ export class Transpiler {
        'library operator part set static sync typedef yield')
           .split(/ /);
 
-  private getLibraryName(nameForTest?: string) {
+  getLibraryName(nameForTest?: string) {
     var fileName = this.getRelativeFileName(nameForTest);
     var parts = fileName.split('/');
     return parts.filter((p) => p.length > 0)
@@ -302,18 +267,6 @@ export class Transpiler {
     }
 
     switch (node.kind) {
-      case ts.SyntaxKind.SourceFile:
-        if (this.options.generateLibraryName) {
-          this.emit('library');
-          this.emit(this.getLibraryName());
-          this.emit(';');
-        }
-        ts.forEachChild(node, this.visit.bind(this));
-        break;
-      case ts.SyntaxKind.EndOfFileToken:
-        ts.forEachChild(node, this.visit.bind(this));
-        break;
-
       case ts.SyntaxKind.NumberKeyword:
         this.emit('num');
         break;
@@ -456,74 +409,6 @@ export class Transpiler {
           this.visit(typeParam.constraint);
         }
         break;
-
-      case ts.SyntaxKind.ImportDeclaration:
-        var importDecl = <ts.ImportDeclaration>node;
-        // TODO(martinprobst): Re-enable once moved to TypeScript.
-        // if (this.isEmptyImport(importDecl)) return;
-        this.emit('import');
-        this.visitExternalModuleReferenceExpr(importDecl.moduleSpecifier);
-        if (importDecl.importClause) {
-          this.visit(importDecl.importClause);
-        } else {
-          this.reportError(importDecl, 'bare import is unsupported');
-        }
-        this.emit(';');
-        break;
-      case ts.SyntaxKind.ImportClause:
-        var importClause = <ts.ImportClause>node;
-        if (importClause.name) this.visitTypeName(importClause.name);
-        if (importClause.namedBindings) {
-          this.visit(importClause.namedBindings);
-        }
-        break;
-      case ts.SyntaxKind.NamespaceImport:
-        var nsImport = <ts.NamespaceImport>node;
-        this.emit('as');
-        this.visitTypeName(nsImport.name);
-        break;
-      case ts.SyntaxKind.NamedImports:
-        this.emit('show');
-        // TODO(martinprobst): Re-enable once moved to TypeScript.
-        // var used = this.filterImports((<ts.NamedImports>node).elements);
-        // if (used.length === 0) {
-        //  this.reportError(node, 'internal error, used imports must not be empty');
-        // }
-        this.visitList((<ts.NamedImports>node).elements);
-        break;
-      case ts.SyntaxKind.NamedExports:
-        this.emit('show');
-        this.visitList((<ts.NamedExports>node).elements);
-        break;
-      case ts.SyntaxKind.ImportSpecifier:
-      case ts.SyntaxKind.ExportSpecifier:
-        var spec = <ts.ImportOrExportSpecifier>node;
-        if (spec.propertyName) this.visitTypeName(spec.propertyName);
-        this.visitTypeName(spec.name);
-        break;
-      case ts.SyntaxKind.ExportDeclaration:
-        var exportDecl = <ts.ExportDeclaration>node;
-        this.emit('export');
-        if (exportDecl.moduleSpecifier) {
-          this.visitExternalModuleReferenceExpr(exportDecl.moduleSpecifier);
-        } else {
-          this.reportError(node, 're-exports must have a module URL (export x from "./y").');
-        }
-        if (exportDecl.exportClause) this.visit(exportDecl.exportClause);
-        this.emit(';');
-        break;
-      case ts.SyntaxKind.ImportEqualsDeclaration:
-        var importEqDecl = <ts.ImportEqualsDeclaration>node;
-        this.emit('import');
-        this.visit(importEqDecl.moduleReference);
-        this.emit('as');
-        this.visitTypeName(importEqDecl.name);
-        this.emit(';');
-        break;
-      case ts.SyntaxKind.ExternalModuleReference:
-        this.visitExternalModuleReferenceExpr((<ts.ExternalModuleReference>node).expression);
-        break;
-
       default:
         this.reportError(node,
             `Unsupported node type ${(<any>ts).SyntaxKind[node.kind]}: ${node.getFullText()}`);
